+++ flipper/hotspot-arcade/scenes/hotspot_arcade_scene_lobby.c
@@ -23,6 +23,8 @@ static const char* game_name(uint8_t g) {
         return "Would You Rather";
     case HA_GAME_SCRAMBLE:
         return "Word Scramble";
+    case HA_GAME_GUESSCOLOR:
+        return "Guess the Color";
     default:
         return "None";
     }